Moved files for blog to correct locations
